requirements.txt
httpx>=0.24.0
python-dotenv>=1.0.0
